# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ODBUS Consumer APIs",
        "description": "L5 Swagger OpenApi description for ODBUS Consumer APIs",
        "version": "0.1"
    },
    "paths": {
...
        "/api/PopularRoutes": {
            "get": {
                "tags": [
                    "PopularRoutes API"
                ],
                "summary": "get all Popular Routes",
                "description": "get all Popular Routes",
                "operationId": "4041a5107390129607bbbf69ced3b3c8",
                "responses": {
                    "200": {
                        "description": "Get all Popular Routes"
                    },
                    "206": {
                        "description": "validation error"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "No record found"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "502": {
                        "description": "Bad gateway"
                    },
                    "503": {
                        "description": "Service unavailable"
                    },
                    "504": {
                        "description": "Gateway timeout"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/TopOperators": {
            "get": {
                "tags": [
                    "TopOperators API"
                ],
                "summary": "get all Top Operators",
                "description": "get all Top Operators",
                "operationId": "c24915a8a9ba6e55266fcef0a7ecf6af",
                "responses": {
                    "200": {
                        "description": "Get all Top Operators"
                    },
                    "206": {
                        "description": "validation error"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "No record found"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "502": {
                        "description": "Bad gateway"
                    },
                    "503": {
                        "description": "Service unavailable"
                    },
                    "504": {
                        "description": "Gateway timeout"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/AllRoutes": {
            "get": {
                "tags": [
                    "All Routes API"
                ],
                "summary": "get all Routes of bus running",
                "description": "get all Routes of bus running",
                "operationId": "a0c65574ff9212f60e88a8a1b0728d5f",
                "responses": {
                    "200": {
                        "description": "Record Fetched Successfully"
                    },
                    "206": {
                        "description": "validation error"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "No record found"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "502": {
                        "description": "Bad gateway"
                    },
                    "503": {
                        "description": "Service unavailable"
                    },
                    "504": {
                        "description": "Gateway timeout"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/AllOperators": {
            "post": {
                "tags": [
                    "All Operators API"
                ],
                "summary": "get all Operators names",
                "description": "get all Operators names",
                "operationId": "d5f2d154f4e53b8d6088f3ff21ad4cf0",
                "parameters": [
                    {
                        "name": "filter params(example:A)",
                        "in": "query",
                        "description": "filter operator names alphabetically(example:A, it will filter opertors starting name with A only)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "paginate",
                        "in": "query",
                        "description": "paginate all operators",
                        "schema": {
                            "type": "integer",
                            "example": "400"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Record Fetched Successfully"
                    },
                    "206": {
                        "description": "validation error"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "No record found"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "502": {
                        "description": "Bad gateway"
                    },
                    "503": {
                        "description": "Service unavailable"
                    },
                    "504": {
                        "description": "Gateway timeout"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/downloadapp": {
            "post": {
                "tags": [
                    "sending sms to download ODBUS App"
                ],
                "summary": "sending sms to download ODBUS App",
                "description": "sending sms to download ODBUS App",
                "operationId": "94f33035b702a34ea4180253bd86c074",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "user mobile number",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Record Fetched Successfully"
                    },
                    "206": {
                        "description": "validation error"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "No record found"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "502": {
                        "description": "Bad gateway"
                    },
                    "503": {
                        "description": "Service unavailable"
                    },
                    "504": {
                        "description": "Gateway timeout"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/OperatorDetails": {
            "get": {
                "tags": [
                    "Operator's Details"
                ],
                "summary": "Operator's Details",
                "description": "Operator's Details",
                "operationId": "f81d1f0dc2025ea4504b7d6e23c183c9",
                "parameters": [
                    {
                        "name": "operator_url",
                        "in": "query",
                        "description": "operator url",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "iconic-travels"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Record Fetched Successfully"
                    },
                    "206": {
                        "description": "validation error"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "No record found"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "502": {
                        "description": "Bad gateway"
                    },
                    "503": {
                        "description": "Service unavailable"
                    },
                    "504": {
                        "description": "Gateway timeout"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "name": "Token based",
                "in": "header",
                "bearerFormat": "AUTH0",
                "scheme": "bearer"
            }
        }
    }
}